+++ src/main/java/com/breakinblocks/nautec/datagen/BlockTagProvider.java
@@ -34,6 +34,7 @@ public class BlockTagProvider extends BlockTagsProvider {
                 POLISHED_PRISMARINE,
                 AQUARINE_STEEL_BLOCK,
                 AQUATIC_CATALYST,
+                SUBMARINE_DOCK,
                 PRESSURE_FORGE,
                 GATEWAY,
                 RESONANCE_CHAMBER,